requirements.txt
discord.py
loguru
pip
GitPython
discord-timers
sqlalchemy
PyMySQL
